refactor(web): move Genres into Discover as sub-tab
- Remove Genres from main navigation (6 → 5 items)
- Add Discover sub-tabs: [Discover] [Genres]
- New routes:
  * /discover (default tab)
  * /discover/genres (genres grid)
  * /discover/genres/{id_or_slug} (genre detail)
- Support both numeric ID and slug in URLs
  * /discover/genres/80 (backward compatible)
  * /discover/genres/jazz (clean URLs)
- Update genre-detail back link to /discover/genres
- Better mobile UX with cleaner navigation

Files changed:
- web/src/routes/discover.rs: Add genres routes
- web/src/routes.rs: Remove genres module
- web/src/lib.rs: Remove genres route registration
- web/templates/discover.html: Add sub-tabs
- web/templates/genre-detail.html: Fix back link
- web/templates/navigation.html: Remove Genres icon
- web/src/routes/genres.rs: Deleted (moved to discover)
