+++ qobuz-player-web/templates/genre-detail.html
@@ -1,9 +1,9 @@
-@defer (base.html; active_page="Genres"; playing_info=playing_info) {
+@defer (base.html; active_page="Discover"; playing_info=playing_info) {
   <div class="flex flex-col gap-8 px-4">
     <div
       class="pt-safe-or-4 sticky top-0 flex max-h-full grow flex-col gap-4 bg-black/80 pb-2 backdrop-blur"
     >
-      <a href="/genres" class="text-blue-500 hover:text-blue-400">
+      <a href="/discover/genres" class="text-blue-500 hover:text-blue-400">
         ← Back to Genres
       </a>
       <h1 class="text-2xl">{{ genre_albums.genre.name }}</h1>
